Autodetect Azure Data Lake Storage
Automatically detect whether the configured Azure Storage account is a
gen1 Azure Blob Storage account or a gen2 Azure Data Lake account.
Switch to the azdatalake client transparently if the account is gen2.

The user can override auto-detection using the azure objstore client
setting azure_storage_account_type; set it to "blob" to force use of a
gen1 client on a gen2 store.